get rid of or document strcat, strncat, strcpy, etc calls. addconn.c: avoid improbable buffer underrun (if confdir is "") addconn.c: correct memory allocation for variable configfile simplify show_one_sr(), fmt_ipsec_sa_established() demux.h, ipsec_doi.c, whackinit.c: size_t should be used where appropriate

+++ programs/pluto/demux.h
@@ -106,9 +106,9 @@ extern void release_any_md(struct msg_digest **mdp);
 typedef stf_status state_transition_fn (struct msg_digest *md);
 
 extern void fmt_ipsec_sa_established(struct state *st,
-				     char *sadetails, int sad_len);
+				     char *sadetails, size_t sad_len);
 extern void fmt_isakmp_sa_established(struct state *st,
-				      char *sadetails, int sad_len);
+				      char *sadetails, size_t sad_len);
 
 extern void free_md_pool(void);
